server/requirements.txt
Flask==1.1.2
flask_cors==3.0.9
neo4j==4.1.1
tmdbsimple==2.6.6
python-dotenv==0.14.0
